refactor: home screen package
