+++ app/src/main/java/com/london/tudee/presentation/screens/home/HomeInteractions.kt
@@ -6,23 +6,23 @@ import com.london.tudee.domain.entities.Task
 import com.london.tudee.presentation.screens.task.TaskModifyInteractions
 
 interface HomeInteractions : TaskModifyInteractions {
-    override fun showDatePicker()
-    override fun hideDatePicker()
     fun showBottomSheet()
-    override fun hideBottomSheet()
-    fun onThemeSwitched(isDarkMode: Boolean)
     fun loadCategories()
     fun validateForm()
     fun clearMessages()
-    fun loadTask(task: Task)
     fun onClickMove()
     fun showTaskDetailsBottomSheet()
     fun hideTaskDetailsBottomSheet()
+    fun loadTask(task: Task)
     fun onEditTask(taskId: Int)
-    override fun updateTitle(title: String)
-    override fun updateDescription(description: String)
+    fun onThemeSwitched(isDarkMode: Boolean)
+    override fun saveTask()
+    override fun showDatePicker()
+    override fun hideDatePicker()
+    override fun hideBottomSheet()
     override fun updateDate(date: Long)
+    override fun updateTitle(title: String)
     override fun updatePriority(priority: Priority)
     override fun updateCategory(category: Category)
-    override fun saveTask()
+    override fun updateDescription(description: String)
 }
