--- app/src/main/java/com/london/tudee/presentation/screens/home/TaskStatusUiState.kt
@@ -1,10 +1,0 @@
-package com.london.tudee.presentation.screens.home
-
-data class TaskStatusUiState(
-    val title: String = "",
-    val subtitle: String = "",
-    val emoji: Int = -1,
-    val tudeePicture: Int = -1,
-)
-
-
